Restore copy and paste functionality.  This was lost when the first responder was changed to a straight UITextField.
There is still a problem that you can't copy and paste with the keyboard hidden
because the "canPerformAction" selector is tied to the first responder (which is
resigned when the keyboard goes away).  To fix this,  the canPerformAction
should be moved to some other object that gets the events forwarded from other
first responders.



git-svn-id: http://mobileterminal.googlecode.com/svn/branches/applesdk/MobileTerminal@454 58fa8ae0-6a36-0410-a7a1-b30e0234c60e