[app] fix the bars so they work on iOS 11 + iPhone X
Seems like iOS 11’s UIToolbar, when the app is linked against SDK 11.0+, eats
touches that aren’t going to button items it’s aware of. Work around it by just
using the toolbar as a background for the other views, rather than adding those
views as subviews to the toolbar.